refactor: streamline tests layout
