+++ test/client/ColumnsPhotoAlbum.spec.tsx
@@ -1,6 +1,6 @@
-import { ColumnsPhotoAlbum } from "../src";
-import { render, renderAndMatchSnapshot } from "./test-utils";
-import photos from "./photos";
+import { ColumnsPhotoAlbum } from "../../src";
+import { render, renderAndMatchSnapshot } from "../test-utils";
+import photos from "../photos";
 
 describe("ColumnsPhotoAlbum", () => {
   it("renders without crashing", () => {
